Moved More Code to Cores
